Extract diagnostics tools; make tool registration reload-safe
Move file_outline, diagnostic_messages, and code_actions into
tools/diagnostics.py. Register tool modules via importlib.import_module
(with a sys.modules pop) so reloading server re-registers tools on the new
mcp instance, which importlib.reload-based tests rely on.
