+++ src/lean_lsp_mcp/tools/__init__.py
@@ -4,8 +4,7 @@ Each submodule defines a thematic group of ``@mcp.tool`` handlers. Importing a
 submodule registers its tools on the shared ``mcp`` instance (defined in
 ``lean_lsp_mcp.server``). Submodules reference shared infrastructure via the
 ``server`` module object so that test monkeypatching keeps working.
+
+Registration is driven by ``server`` (see ``TOOL_MODULES`` there), which imports
+these submodules after the core is defined.
 """
-
-from lean_lsp_mcp.tools import build, goals, widgets  # noqa: F401
-
-__all__ = ["build", "goals", "widgets"]
